Allow to convert multiple files
Enhance command line syntax and introduce a few more options.
Multiple input files can now be specified on the command line.

For backwards compatibility the following behavior is implemented:
If exactly two parameters are specified the first is treated as input,
the other as output. If only one parameter is specified it is
treated as input and the output derived from the filename with the
extension changed to html.

If no output filename is specified as argument the default name
NUnitOutput.html is used.

New options:
-f/--force    to overwrite existing output file
-v/--verbose  to output status messages
-o/--output   to specify output filename